breaking: rename pymatviz/(structure_viz->structure)
