+++ assets/scripts/structure/structure_2d_plotly.py
