+++ assets/scripts/structure/structure_3d_plotly.py
