+++ pymatviz/structure/__init__.py
@@ -1,0 +1,5 @@
+"""2D and 3D plots of Structures."""
+
+from __future__ import annotations
+
+from pymatviz.structure.plotly import structure_2d_plotly, structure_3d_plotly
