+++ pymatviz/structure/helpers.py
