--- pymatviz/structure_viz/__init__.py
@@ -1,5 +1,0 @@
-"""2D and 3D plots of Structures."""
-
-from __future__ import annotations
-
-from pymatviz.structure_viz.plotly import structure_2d_plotly, structure_3d_plotly
