+++ readme.md
@@ -140,13 +140,13 @@ On the roadmap but no ETA yet.
 
 ## Structure
 
-See [`pymatviz/structure_viz/plotly.py`](pymatviz/structure_viz/plotly.py).
+See [`pymatviz/structure/plotly.py`](pymatviz/structure/plotly.py).
 
-|                                 [`structure_3d_plotly(hea_structure)`](pymatviz/structure_viz/plotly.py#L318)                                 | [`structure_3d_plotly(lco_supercell)`](pymatviz/structure_viz/plotly.py#L318) [![fig-icon]](assets/scripts/structure_viz/structure_3d_plotly.py) |
-| :-------------------------------------------------------------------------------------------------------------------------------------------: | :----------------------------------------------------------------------------------------------------------------------------------------------: |
-|                                                          ![hea-structure-3d-plotly]                                                           |                                                            ![lco-structure-3d-plotly]                                                            |
-| [`structure_2d_plotly(six_structs)`](pymatviz/structure_viz/plotly.py#L42) [![fig-icon]](assets/scripts/structure_viz/structure_2d_plotly.py) |  [`structure_3d_plotly(six_structs)`](pymatviz/structure_viz/plotly.py#L318) [![fig-icon]](assets/scripts/structure_viz/structure_3d_plotly.py)  |
-|                                                   ![matbench-phonons-structures-2d-plotly]                                                    |                                                     ![matbench-phonons-structures-3d-plotly]                                                     |
+|                               [`structure_3d_plotly(hea_structure)`](pymatviz/structure/plotly.py#L318)                               | [`structure_3d_plotly(lco_supercell)`](pymatviz/structure/plotly.py#L318) [![fig-icon]](assets/scripts/structure/structure_3d_plotly.py) |
+| :-----------------------------------------------------------------------------------------------------------------------------------: | :--------------------------------------------------------------------------------------------------------------------------------------: |
+|                                                      ![hea-structure-3d-plotly]                                                       |                                                        ![lco-structure-3d-plotly]                                                        |
+| [`structure_2d_plotly(six_structs)`](pymatviz/structure/plotly.py#L42) [![fig-icon]](assets/scripts/structure/structure_2d_plotly.py) |  [`structure_3d_plotly(six_structs)`](pymatviz/structure/plotly.py#L318) [![fig-icon]](assets/scripts/structure/structure_3d_plotly.py)  |
+|                                               ![matbench-phonons-structures-2d-plotly]                                                |                                                 ![matbench-phonons-structures-3d-plotly]                                                 |
 
 [matbench-phonons-structures-2d-plotly]: assets/svg/matbench-phonons-structures-2d-plotly.svg
 [matbench-phonons-structures-3d-plotly]: assets/svg/matbench-phonons-structures-3d-plotly.svg
